Upload of this month's changes
Lots of changes- kept forgetting to upload so it's going to just all be
in one massive commit. Some extra commits later while I work on a few
features.

Coding:
-Game Manager
-Score System
-Win / Loss Conditions
-Audio Manager
-Fixed Hammer Swinging Incorrectly
-Fixed Ranged Enemy Not Rotating Properly
-Added Drop-In System

UI:
-Added UI

Graphics:
-Added Player Projectiles
-Added / Animated Helicopter

Audio:
-Added Weapon Sound Effects
-Added a Few Public Domain Soundtracks
-Added Enemy Death Sounds

I will add more to the notes if I remember, as well as fixing anything
up that is still broken / needs to be added.